fix:db

+++ database.py
@@ -1,6 +1,6 @@
 # coding: utf-8
 import os
-from sqlalchemy import Column, INTEGER, TEXT, BOOLEAN, DATETIME, TIMESTAMP, create_engine
+from sqlalchemy import Column, INTEGER, BIGINT, TEXT, BOOLEAN, DATETIME, TIMESTAMP, create_engine
 from sqlalchemy.orm import sessionmaker
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.pool import StaticPool
@@ -30,7 +30,7 @@ class Message(Base):
     __tablename__ = 'message'
 
     _id = Column(INTEGER, primary_key=True)
-    id = Column(INTEGER)
+    id = Column(BIGINT)
     link = Column(TEXT)
     type = Column(TEXT)  # 文本、图像、视频、音频、语音
     category = Column(TEXT)  # 分类
@@ -40,14 +40,14 @@ class Message(Base):
     audio = Column(TEXT)
     voice = Column(TEXT)
     date = Column(TIMESTAMP)
-    from_id = Column(INTEGER)
-    from_chat = Column(INTEGER)
+    from_id = Column(BIGINT)
+    from_chat = Column(BIGINT)
 
 
 class User(Base):
     __tablename__ = 'user'
 
-    id = Column(INTEGER, primary_key=True)
+    id = Column(BIGINT, primary_key=True)
     fullname = Column(TEXT)
     username = Column(TEXT)
 
@@ -56,16 +56,16 @@ class UserAlias(Base):
     __tablename__ = 'user_alias'
 
     id = Column(INTEGER, primary_key=True)
-    user_id = Column(INTEGER)  # The actual user ID this alias refers to
+    user_id = Column(BIGINT)  # The actual user ID this alias refers to
     alias = Column(TEXT)  # The alias name
-    created_by = Column(INTEGER)  # The user ID who created this alias
+    created_by = Column(BIGINT)  # The user ID who created this alias
     created_at = Column(TIMESTAMP)  # When this alias was created
 
 
 class Chat(Base):
     __tablename__ = 'chat'
 
-    id = Column(INTEGER, primary_key=True)
+    id = Column(BIGINT, primary_key=True)
     title = Column(TEXT)
     enable = Column(BOOLEAN)
